style: creation of the styles of the login and registration screens and separation of the route file
