+++ home/src/app/account/create-account/create-account.component.css
@@ -1,0 +1,3 @@
+button {
+  margin-top: 25px;
+}
