+++ home/src/app/app-routing.module.ts
@@ -1,0 +1,34 @@
+import { AuthGuard } from './account/shared/auth.guard';
+import { CreateAccountComponent } from './account/create-account/create-account.component';
+import { LoginComponent } from './account/login/login.component';
+import { AuthenticationComponent } from './layout/authentication/authentication.component';
+import { HomeComponent } from './layout/home/home.component';
+import { ProductListComponent } from './products/product-list/product-list.component';
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+const appRoutes: Routes = [
+  {
+    path: '',
+    component: HomeComponent,
+    children: [
+      { path: '', component: ProductListComponent }
+    ],
+    canActivate: [AuthGuard]
+  },
+  {
+    path: '',
+    component: AuthenticationComponent,
+    children: [
+      { path: '', redirectTo: 'login', pathMatch: 'full' },
+      { path: 'login', component: LoginComponent },
+      { path: 'create-account', component: CreateAccountComponent }
+    ]
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forRoot(appRoutes)],
+  exports: [RouterModule]
+})
+export class AppRoutingModule { }
